requirements.txt
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
transformers
torch
pillow
pydantic
reverse_geocoder
numpy